requirements.txt
torch==1.13.1 
torchvision==0.14.1
numpy==1.24.2
Pillow==9.5.0
pandas==2.0.0
opencv-python==4.7.0.72
scikit-learn==1.2.2
pygame==2.3.0
tqdm~=4.65.0
tensorflow~=2.12.0
